+++ components/ajouter-cas-types/form/cas-types-revenus.jsx
@@ -7,8 +7,8 @@ import PropTypes from "prop-types";
 import { Fragment } from "react";
 import { Field } from "react-final-form";
 
-import { formatNumber } from "../../utils";
-import generateRevenusMensuel from "../../utils/maths/generate-revenus-mensuel";
+import { formatNumber } from "../../common";
+import generateRevenusMensuel from "../../common/utils/maths/generate-revenus-mensuel";
 
 const REVENUS_MENSUEL = generateRevenusMensuel(500);
 const selectOptions = REVENUS_MENSUEL.map((value) => {
