+++ components/ir/cartes-impact/carte-etat/bar-chart.jsx
