+++ components/ir/cartes-impact/carte-etat/bar-chart.module.scss
