+++ components/ir/cartes-impact/carte-etat/carte-etat-component.tsx
@@ -16,10 +16,10 @@ import { Fragment, PureComponent } from "react";
 // eslint-disable-next-line no-unused-vars
 import { connect, ConnectedProps } from "react-redux";
 
-import { fetchSimPop, simulateCasTypes } from "../../redux/actions";
+import { fetchSimPop, simulateCasTypes } from "../../../../redux/actions";
 // eslint-disable-next-line no-unused-vars
-import { RootState } from "../../redux/reducers";
-import { Card } from "../common";
+import { RootState } from "../../../../redux/reducers";
+import { Card } from "../../../common";
 import BarChart from "./bar-chart";
 import styles2 from "./carte-etat-component.module.scss";
 import SimpopTableurInfosDeciles from "./simpop-tableur-infos-deciles";
