+++ components/ir/cartes-impact/carte-etat/simpop-tableur-infos-deciles.module.scss
